feat(abstract-lightning): add codecs for more lightning apis
get payment, get transaction
and codec refactor and cleanup
change signerIP and expressIP to expressHost for more flexibility

Ticket: BTC-1837

BREAKING CHANGE: Lightning v2
