+++ modules/abstract-lightning/src/index.ts
@@ -1,3 +1,4 @@
 export * from './abstractLightningCoin';
 export * from './wallet';
 export * from './lightning';
+export * from './codecs';
