+++ modules/abstract-lightning/src/lightning/index.ts
@@ -1,5 +1,3 @@
 export * from './signableJson';
 export * from './signature';
 export * from './lightningUtils';
-export * from './codecs';
-export * from './wpCodecs';
