+++ modules/abstract-lightning/src/wallet/lightning.ts
@@ -1,16 +1,15 @@
 import * as sdkcore from '@bitgo/sdk-core';
 import * as t from 'io-ts';
+import { createMessageSignature, unwrapLightningCoinSpecific } from '../lightning';
 import {
-  createMessageSignature,
+  CreateInvoiceBody,
+  Invoice,
+  InvoiceInfo,
+  InvoiceQuery,
   LightningAuthKeychain,
   LightningKeychain,
-  unwrapLightningCoinSpecific,
   UpdateLightningWalletSignedRequest,
-  CreateInvoiceRequest,
-  GetInvoicesQuery,
-  LightningInvoice,
-  InvoiceInfo,
-} from '../lightning';
+} from '../codecs';
 
 export interface ILightningWallet {
   /**
@@ -19,9 +18,9 @@ export interface ILightningWallet {
    * @param {bigint} params.valueMsat The value of the invoice in millisatoshis
    * @param {string} [params.memo] A memo or description for the invoice
    * @param {number} [params.expiry] The expiry time of the invoice in seconds
-   * @returns {Promise<LightningInvoice>} A promise that resolves to the created invoice
+   * @returns {Promise<Invoice>} A promise that resolves to the created invoice
    */
-  createInvoice(params: CreateInvoiceRequest): Promise<LightningInvoice>;
+  createInvoice(params: CreateInvoiceBody): Promise<Invoice>;
 
   /**
    * Lists current lightning invoices
@@ -34,7 +33,7 @@ export interface ILightningWallet {
    * @param {Date} [params.startDate] The start date for the query
    * @param {Date} [params.endDate] The end date for the query
    */
-  listInvoices(params: GetInvoicesQuery): Promise<InvoiceInfo[]>;
+  listInvoices(params: InvoiceQuery): Promise<InvoiceInfo[]>;
 
   /**
    * Pay a lightning invoice
@@ -60,7 +59,7 @@ export interface ILightningWallet {
    *     Encrypted with ECDH secret key from private key of wallet's user auth key and public key of lightning service.
    *   - `encryptedSignerAdminMacaroon` (optional): Generated when initializing the wallet of the signer node.
    *     Encrypted with client's wallet passphrase.
-   *   - `signerIp` (optional): The IP address of the Lightning signer node.
+   *   - `signerHost` (optional): The host address of the Lightning signer node.
    *   - `encryptedSignerTlsKey` (optional): The wallet passphrase encrypted TLS key of the signer.
    *   - `signerTlsCert` (optional): The TLS certificate of the signer.
    *   - `watchOnlyAccounts` (optional): These are the accounts used to initialize the watch-only wallet.
@@ -81,12 +80,12 @@ export class SelfCustodialLightningWallet implements ILightningWallet {
     this.wallet = wallet;
   }
 
-  async createInvoice(params: CreateInvoiceRequest): Promise<LightningInvoice> {
+  async createInvoice(params: CreateInvoiceBody): Promise<Invoice> {
     const createInvoiceResponse = await this.wallet.bitgo
       .post(this.wallet.baseCoin.url(`/wallet/${this.wallet.id()}/lightning/invoice`))
-      .send(CreateInvoiceRequest.encode(params))
+      .send(CreateInvoiceBody.encode(params))
       .result();
-    return sdkcore.decodeOrElse(LightningInvoice.name, LightningInvoice, createInvoiceResponse, (error) => {
+    return sdkcore.decodeOrElse(Invoice.name, Invoice, createInvoiceResponse, (error) => {
       // DON'T throw errors from decodeOrElse. It could leak sensitive information.
       throw new Error(`Invalid create invoice response ${error}`);
     });
@@ -96,11 +95,11 @@ export class SelfCustodialLightningWallet implements ILightningWallet {
     throw new Error('Method not implemented.');
   }
 
-  async listInvoices(params: GetInvoicesQuery): Promise<InvoiceInfo[]> {
+  async listInvoices(params: InvoiceQuery): Promise<InvoiceInfo[]> {
     const returnCodec = t.array(InvoiceInfo);
     const createInvoiceResponse = await this.wallet.bitgo
       .get(this.wallet.baseCoin.url(`/wallet/${this.wallet.id()}/lightning/invoice`))
-      .query(GetInvoicesQuery.encode(params))
+      .query(InvoiceQuery.encode(params))
       .result();
     return sdkcore.decodeOrElse(returnCodec.name, returnCodec, createInvoiceResponse, (error) => {
       throw new Error(`Invalid list invoices response ${error}`);
